refactor : community form, dto
