+++ src/main/java/damon/backend/dto/request/community/CommunityCommentCreateForm.java
@@ -1,6 +1,6 @@
 package damon.backend.dto.request.community;
 
-import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
@@ -10,6 +10,6 @@ public class CommunityCommentCreateForm {
     @NotNull
     private Long communityId;
 
-    @NotBlank(message = "댓글 내용을 입력해주세요.")
+    @NotEmpty(message = "내용을 입력해주세요.")
     private String content;
 }
