+++ src/main/java/damon/backend/dto/request/community/CommunityUpdateForm.java
@@ -1,6 +1,6 @@
 package damon.backend.dto.request.community;
 
-import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
@@ -12,10 +12,10 @@ public class CommunityUpdateForm {
     @NotNull
     private Long communityId;
 
-    @NotBlank(message = "제목을 입력해주세요.")
+    @NotEmpty(message = "제목을 입력해주세요.")
     private String title;
 
-    @NotBlank(message = "내용을 입력해주세요.")
+    @NotEmpty(message = "내용을 입력해주세요.")
     private String content;
 
     private List<String> images;
