+++ src/main/java/damon/backend/dto/response/community/CommunityDetailDTO.java
@@ -1,10 +1,11 @@
 package damon.backend.dto.response.community;
 
-import damon.backend.entity.Community;
+import damon.backend.entity.community.Community;
 import damon.backend.enums.CommunityType;
 import lombok.Data;
 
 import java.time.LocalDateTime;
+import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -25,6 +26,7 @@ public class CommunityDetailDTO {
     private List<CommunityCommentDTO> comments;
 
     public CommunityDetailDTO(Community community) {
+
         this.communityId = community.getCommunityId();
         this.memberId = community.getMember().getId();
         this.memberName = community.getMember().getNickname();
@@ -41,7 +43,7 @@ public class CommunityDetailDTO {
                 .map(CommunityLikeDTO::new)
                 .collect(Collectors.toList());
 
-        this.comments = community.getComments()
+        this.comments = community.getComments().reversed() // 역순으로 정렬
                 .stream()
                 .map(CommunityCommentDTO::new)
                 .collect(Collectors.toList());
