+++ src/main/java/damon/backend/dto/response/community/CommunityLikeDTO.java
@@ -1,6 +1,6 @@
 package damon.backend.dto.response.community;
 
-import damon.backend.entity.CommunityLike;
+import damon.backend.entity.community.CommunityLike;
 import lombok.Data;
 
 @Data
